Adding GetOutputDir method to test_support library.
The unittest is not ideal for this, but I would have to use similar code as the implementation of the GetOutputDir in order to verify that it actually runs, so it wouldn't make much sense with a test like that.

It compiles and runs on Linux, Win and Mac. The folder gets created and is writeable from other tests.

I have tried using the GetOutputDir from another project that writes output files and it works as intended on all platforms.

Review URL: http://webrtc-codereview.appspot.com/270001

git-svn-id: https://webrtc.googlecode.com/svn@906 4adac7df-926f-26a2-2b94-8c16560cd09d