Improve the mixing saturation protection scheme.
A single participant is not processed at all. With multiple
participants, we divide-by-2 as before when mixing. Afterwards,
the mixed signal is limited by the AGC to -7 dBFS and then doubled to
restore the original level.

This preserves the level while guaranteeing good saturation protection.

Add a test to voe_auto_test. Hijack and improve the existing mixing test
for this.

TEST=voe_auto_test, voe_cmd_test

Review URL: http://webrtc-codereview.appspot.com/241013

git-svn-id: https://webrtc.googlecode.com/svn/trunk@920 4adac7df-926f-26a2-2b94-8c16560cd09d