Integration test that tracks dropped frames and compares video output.
The recorded frame timestamps are used to modify the output video on a frame-per-frame so it can be compared with the reference video using PSNR. This code will make it possible to use vie_auto_test for full stack comparisons with network interference and similar interesting simulations.

There's some refactoring done in vie_comparison_test.cc to make it fit to the new test.

Compiled and executed in Debug+Release on Linux, Mac and Windows.

BUG=
TEST=vie_auto_test --automated --gtest_filter=ViEVideoVerificationTest.*

Review URL: http://webrtc-codereview.appspot.com/320002

git-svn-id: https://webrtc.googlecode.com/svn@1269 4adac7df-926f-26a2-2b94-8c16560cd09d