Properly count anonymous mixing participants.
When _amountOfMixableParticipants == 1, we skip mixing and saturation
protection. Without this fix, an anonymous participant would only be
properly counted if it was the last added.

For example, if an anonymous participant was added first, followed by
a regular participant, _amoutOfMixableParticipants would == 1 and the
regular participant would not be mixed.

BUG=issue209
TEST=New test added to voe_auto_test to verify, and used voe_cmd_test.

Review URL: https://webrtc-codereview.appspot.com/367006

git-svn-id: https://webrtc.googlecode.com/svn/trunk@1551 4adac7df-926f-26a2-2b94-8c16560cd09d