This removes the knowledge of frame completeness from the FEC decoder. Therefore, with this change a recovered packet is only considered old, and will be removed, if more than 48 recovered packets are stored.
Packets are immediately reconstructed and sent to the jitter
buffer. Before this CL packets were processed on a frame-by-frame
basis, and all packets belonging to a frame was sent to the
jitter buffer at the same time.

The number of FEC packets is also limited to 48. An FEC packet is
removed if all protected packets have been recovered or if the
FEC packet is considered old.

Lot's of tests added.

Patch-set 2:
- Fixed rtp_fec_unittest.cc to work with the new reconstruction.
- Added reference counting of Packet to be able to keep references to them from FecPacket between different reconstruction runs.
- Rewrote the packet search to use std::set_intersection.

BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/379005

git-svn-id: https://webrtc.googlecode.com/svn@1651 4adac7df-926f-26a2-2b94-8c16560cd09d