Rolled back r2177 since it breaks vie_auto_test.
I'm not sure what is the right thing to do here. That would probably be to call Release() the right amount of times and ensuring that the last call returns 0 (e.g. all references have been released), but I'll leave it up to the CL author to investigate that.

TBR=elham@webrtc.org
BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/569005

git-svn-id: https://webrtc.googlecode.com/svn/trunk@2183 4adac7df-926f-26a2-2b94-8c16560cd09d