Hi Tina,
I have uploaded this patch for your review. I have done an extensive test to be sure that removing of tables does not create any problem. 

The test file, is called test_lpc.c which requires a hack to standard iSAC. The test computes LPC coefficients, then encodes and decodes with old and new (size-reduced) tables. It compares the results is all steps. I have ran the test over large set of files, more then 51 hours of audio, and there was no error. 

I tried to do no formatting so the review to be easier, but I know it can be a tricky CL. Hopefully, the test file helps you to be more confident on the CL. 

Thanks,... Turaj  

In this change list the LPC tables associated with mode 1 & 2 are remoded, and necessary cahnges are made to other files. 

The only model allowed is model number 0. Therefore, this CL breaks compatibility with iSAC released prior to 2.4.3. To avoid changing the bit-stream, we still keep the model number in the bit-stream. 

entropy_coding.c is cleaned up, especially encoding of LAR had KLT transform of LPC gains which are removed now. 
Review URL: https://webrtc-codereview.appspot.com/548004

git-svn-id: https://webrtc.googlecode.com/svn/trunk@2186 4adac7df-926f-26a2-2b94-8c16560cd09d