Enable aecm neon optimized code in android build
Description:
This CL is a follow up of http://review.webrtc.org/575008/ and r2241.
Because of r2243, r2241 is messed up and reverted, I'm going to commit it again.
This is exact same to the last patch in 575008, just want to inform you and have
your double check before I commit it.

The original description
This cl enables cpu detection and aecm optimized code in android build.

BUG=
TEST=build on all platforms
Review URL: https://webrtc-codereview.appspot.com/568006

git-svn-id: https://webrtc.googlecode.com/svn/trunk@2259 4adac7df-926f-26a2-2b94-8c16560cd09d