I haven't done any refactoring here.
Resolve coverity warnings.

14305.

The warning is not really valid. The 'decode' function should be called with a 'mode' variable, where inside the function it is assumed that mode is either zero or one. If mode is taking other values some varibles are used uninitialized. However, this is an internal function and it is always called with either ZERO or ONE. Therefore, the code operates correctly. I made small changes as I beleive it is a bit nicer way. 

In ACM:
- Conditions on 'mode' is changed.


Tested with trybots.

BUG=None
TEST=None
Review URL: https://webrtc-codereview.appspot.com/564014

git-svn-id: https://webrtc.googlecode.com/svn/trunk@2297 4adac7df-926f-26a2-2b94-8c16560cd09d