Allow NetEQ to use real packet durations.
This adds a FuncDurationEst to each codec instance which estimates
the duration of a packet given the packet contents. By default, the
duration of the previous packet is used (which is what is currently
assumed to be the duration of all future packets). This patch also
provides an initial implementation of this function for G.711, which
returns the actual number of samples in the packet.