Adding AUDIO application as default for Opus stereo
The Opus audio codec targets applications for pure conversations as well as other types of audio (e.g. music), and there are two different settings to use for this (VoIP and AUDIO). In the current implementation of Opus in WebRTC we use VoIP only.

I this CL I have changed default setting to AUDIO in the case of stereo, and kept VoIP as default in case of mono.

Next step is to add an API to choose application mode.

BUG=issue1239

Review URL: https://webrtc-codereview.appspot.com/1007006

git-svn-id: https://webrtc.googlecode.com/svn/trunk@3319 4adac7df-926f-26a2-2b94-8c16560cd09d