Enable indefinitely running vie_auto_test option
When doing test automation, the prompt in vie_auto_test is not working as expected on Windows when the test is run from a Buildbot. As soon a prompt is presented to the test runner, vie_auto_test exits, assuming the user pressed Ctrl-D.

By adding a third option for the Stop/Modify call prompt that allows running the call indefinitely (and making that the default), no prompt is displayed when the --auto_custom_call flag is used.

BUG=none
TEST=Execution with vie_auto_test.exe --auto_custom_call --override "Enter destination IP.=192.168.3.11" and by running vie_auto_test in interactive mode.
+ Trybots passing.

Review URL: https://webrtc-codereview.appspot.com/1099004

git-svn-id: https://webrtc.googlecode.com/svn/trunk@3478 4adac7df-926f-26a2-2b94-8c16560cd09d